refactor(syncing): move peer to separate file & execute synchronously PBFT block on syncing
