+++ .gitignore
@@ -21,3 +21,6 @@ clang-tidy
 
 # header with version of taraxad (it generated by cmake)
 src/config/version.hpp
+
+# python test related folders
+__pycache__/
